refactor(meetings): the create bubbles function is in meetings module
